+++ packages/js/src/make_script_data/_script_data_type.ts
@@ -2,23 +2,28 @@ import z from 'zod';
 import { SCRIPT_LIST } from '../utils/lang_list';
 // import type { OutputScriptData } from './output_script_data_schema';
 
+const MAX_U8_INT = 255;
+
+const max_int = z.int().max(MAX_U8_INT);
+const max_int_min_0 = max_int.min(0);
+
 const commonScriptDataSchema = z.strictObject({
   /**  [krama_key: string, list_arr_ref: number | null] */
-  krama_text_arr: z.tuple([z.string(), z.int().nullable()]).array(),
-  krama_text_arr_index: z.int().array(),
+  krama_text_arr: z.tuple([z.string(), max_int_min_0.nullable()]).array(),
+  krama_text_arr_index: max_int_min_0.array(),
   text_to_krama_map: z
     .tuple([
       z.string(), // text
       z.strictObject({
         next: z.string().array().nullable().optional(),
-        krama: z.int().array().nullable().optional(),
-        fallback_list_ref: z.int().nullable().optional()
+        krama: max_int.array().nullable().optional(),
+        fallback_list_ref: max_int_min_0.nullable().optional()
       })
     ])
     .array(),
   list: z
     .strictObject({
-      krama_ref: z.int().array(),
+      krama_ref: max_int_min_0.array(),
       type: z.literal('anya')
     })
     .array(),
@@ -27,23 +32,23 @@ const commonScriptDataSchema = z.strictObject({
       z.string(), // text
       z.strictObject({
         next: z.string().array().nullable().optional(),
-        krama: z.int().array().nullable().optional(),
-        custom_back_ref: z.int().nullable().optional(),
-        fallback_list_ref: z.int().nullable().optional()
+        krama: max_int.array().nullable().optional(),
+        custom_back_ref: max_int_min_0.nullable().optional(),
+        fallback_list_ref: max_int_min_0.nullable().optional()
       })
     ])
     .array(),
   custom_script_chars_arr: z
     .tuple([
       z.string(), // text
-      z.int().nullable(),
-      z.int().nullable()
+      max_int_min_0.nullable(),
+      max_int_min_0.nullable()
     ])
     .array(),
 
   // input script origin attributes
   script_name: z.enum(SCRIPT_LIST),
-  script_id: z.int(),
+  script_id: max_int_min_0,
   script_type: z.enum(['brahmic', 'other']),
   schwa_character: z.string()
 });
@@ -62,14 +67,14 @@ const brahmicScriptDataSchema = commonScriptDataSchema
       .discriminatedUnion('type', [
         z.strictObject({
           type: z.enum(['anya', 'vyanjana', 'mAtrA']),
-          krama_ref: z.int().array()
+          krama_ref: max_int_min_0.array()
         }),
         z.strictObject({
           type: z.literal('svara'),
-          krama_ref: z.int().array(),
+          krama_ref: max_int_min_0.array(),
           // if there is krama_ref for svara in brahmic type then there also will
           // be mAtrA ref
-          mAtrA_krama_ref: z.int().array()
+          mAtrA_krama_ref: max_int_min_0.array()
         })
       ])
       .array()